requirements.txt
kafka-python==2.0.2
pyspark==3.4.0
psycopg2-binary==2.9.6
faker==18.9.0
pandas==2.0.2